Use a white square emoji for blanks
If the park uses a mixture of regular spaces and emoji the alignment
will not be preserved nicely in many applications, especially web
browsers, because the size of the regular space is not a match for
the size of emojis used.

Emoji themselves are not all guaranteed to be the same size in all
fonts, but at if we use a 'white large square' it appears to be a
more reliable size match for the rest of the emoji.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>